+++ camstack/viewermains/apapane.py
@@ -1,42 +1,27 @@
 #!/usr/bin/env python
 
-# ------------------------ #
-#    ___      _ _ _        #
-#   / _ \__ _| (_) | __ _  #
-#  / /_)/ _` | | | |/ _` | #
-# / ___/ (_| | | | | (_| | #
-# \/    \__,_|_|_|_|\__,_| #
-#                          #
-# ------------------------ #
-
-import signal
-
-
-def print_linenum(signum, frame):
-    print("Currently at line", frame.f_lineno, 'in file',
-          frame.f_code.co_filename)
-
-
-signal.signal(signal.SIGTTIN, print_linenum)
+# ------------------------------------------ #
+#    _                                       #
+#   /_\  _ __   __ _ _ __   __ _ _ __   ___  #
+#  //_\\| '_ \ / _` | '_ \ / _` | '_ \ / _ \ #
+# /  _  \ |_) | (_| | |_) | (_| | | | |  __/ #
+# \_/ \_/ .__/ \__,_| .__/ \__,_|_| |_|\___| #
+#       |_|         |_|                      #
+# ------------------------------------------ #
 
 import os, sys
 
 _CORES = os.sched_getaffinity(0)  # Go around pygame import
 
-signal.signal(signal.SIGTTIN, print_linenum)
-
-import pygame
+import pygame, sys
 from pygame.locals import *
 
 # Pygame import (on AMD Epyc) make affinity drop to CPU 0 only !
 os.sched_setaffinity(0, _CORES)  # Fix the CPU affinity
 
 import numpy as np
-
-from matplotlib import cm
-
+import matplotlib.cm as cm
 import struct
-
 from PIL import Image
 import time
 import math as m
@@ -45,8 +30,6 @@ import datetime as dt
 from astropy.io import fits as pf
 import subprocess
 
-import scxconf
-
 from scxkw.config import REDIS_DB_HOST, REDIS_DB_PORT
 from scxkw.redisutil.typed_db import Redis
 
@@ -54,9 +37,9 @@ from pyMilk.interfacing.isio_shmlib import SHM
 
 import camstack.viewertools.viewer_common as cvc
 
-home = os.getenv('HOME')  # Expected /home/scexao
-conf_dir = home + "/conf/palila_aux/"
-sys.path.append(home + '/src/lib/python/')
+HOME = os.getenv('HOME')  # Expected /home/scexao
+conf_dir = HOME + "/conf/apapane_aux/"
+sys.path.append(HOME + '/src/lib/python/')
 
 MILK_SHM_DIR = os.getenv(
         'MILK_SHM_DIR')  # Expected /tmp <- MULTIVERSE FIXING NEEDED
@@ -71,14 +54,13 @@ ONES_NODIM = np.array(1., dtype=np.float32)
 # ------------------------------------------------------------------
 from camstack.core import tmux as tmuxlib
 
+
 # ------------------------------------------------------------------
 #             short hands for shared memory data access
 # ------------------------------------------------------------------
-
-
 def get_img_data(*args, **kwargs):
-    # Arguments: bias, badpixmap, subt_ref, ref, lin_scale, clean, check
-    return cvc.get_img_data(cam, cvc.CREDWHAT.TWO, *args, **kwargs)
+    # Arguments: bias, badpixmap, subt_ref, ref, line_scale, clean, check
+    return cvc.get_img_data(cam, cvc.CREDWHAT.ONE, *args, **kwargs)
 
 
 # ------------------------------------------------------------------
@@ -97,44 +79,64 @@ def ave_img_data(nave, *args, **kwargs):
 
 
 def arr2im(arr, vmin=0., vmax=10000.0, pwr=1.0, subt_ref=False, lin_scale=True,
-           pos=[0, 0]):
-
+           pos=[0, 0], pdi=False):
     (ysizeim, xsizeim) = arr.shape
     ymin = xmin = 0
     ymax = ysizeim
     xmax = xsizeim
-    if z2 != 1:
-        ymin = max(256 + round(pos[1] + cor[1] - crop[2] - ysizeim / 2. / z2),
-                   0)
-        ymax = min(256 + round(pos[1] + cor[1] - crop[2] + ysizeim / 2. / z2),
-                   ysizeim)
-        xmin = max(320 + round(pos[0] + cor[0] - crop[0] - xsizeim / 2. / z2),
-                   0)
-        xmax = min(320 + round(pos[0] + cor[0] - crop[0] + xsizeim / 2. / z2),
-                   xsizeim)
-        if ymin == 0:
-            ymax = round(ysizeim / float(z2))
-        elif ymax == ysizeim:
-            ymin = ysizeim - round(ysizeim / float(z2))
-        if xmin == 0:
-            xmax = round(xsizeim / float(z2))
-        elif xmax == xsizeim:
-            xmin = xsizeim - round(xsizeim / float(z2))
-        ymin, ymax = int(ymin), int(ymax)
-        xmin, xmax = int(xmin), int(xmax)
-        arr2 = arr[ymin:ymax, xmin:xmax].astype(np.float32)
-    else:
-        arr2 = arr.astype(np.float32)
 
-    if not lin_scale:
-        lmin = np.percentile(arr2, 0.8)
-        arr2 -= lmin
-        mask = arr2 > 0
-        arr2 *= mask
+    if pdi and (xsizeim // z2) >= 160:
+        pdi = False  # image is not zoomed enough to do a dual-center crop
 
+    if not pdi:
+        if z2 != 1:
+            ymin = max(
+                    128 + round(pos[1] + cor[1] - crop[2] - ysizeim / 2. / z2),
+                    0)
+            ymax = min(
+                    128 + round(pos[1] + cor[1] - crop[2] + ysizeim / 2. / z2),
+                    ysizeim)
+            xmin = max(
+                    160 + round(pos[0] + cor[0] - crop[0] - xsizeim / 2. / z2),
+                    0)
+            xmax = min(
+                    160 + round(pos[0] + cor[0] - crop[0] + xsizeim / 2. / z2),
+                    xsizeim)
+            if ymin == 0:
+                ymax = round(ysizeim / float(z2))
+            elif ymax == ysizeim:
+                ymin = ysizeim - round(ysizeim / float(z2))
+            if xmin == 0:
+                xmax = round(xsizeim / float(z2))
+            elif xmax == xsizeim:
+                xmin = xsizeim - round(xsizeim / float(z2))
+            ymin, ymax = int(ymin), int(ymax)
+            xmin, xmax = int(xmin), int(xmax)
+            arr2 = arr[ymin:ymax, xmin:xmax].astype(np.float32)
+        else:
+            arr2 = arr.astype(np.float32)
+    else:  # if pdi, assuming centers = middle -/+ 40 of the 1st axis and middle of 2nd axis
+        cen_row = ymax // 2
+        cen_cols = (xmax // 2 - 37, xmax // 2 + 37)
+        half_size = int(round(ymax / z2 / 2))
+        quart_size = int(round(ymax / z2 / 4))
+        arr2 = np.c_[arr[cen_row - half_size:cen_row + half_size,
+                         cen_cols[0] - quart_size:cen_cols[0] + quart_size],
+                     arr[cen_row - half_size:cen_row + half_size,
+                         cen_cols[1] - quart_size:cen_cols[1] +
+                         quart_size]].astype(np.float32)
+
+    lmin = np.percentile(arr2, 0.1)
+    arr2 -= lmin
+    mask = arr2 > 0
+    arr2 *= mask
+    lmax = np.percentile(arr2, 99.95)
+    mask = arr2 < lmax
+    arr2 *= mask
+    arr2 += (1 - mask) * lmax
     arr3 = arr2**pwr
-    mmin, mmax = arr3[1:].min(), arr3[1:].max(
-    )  # IGNORE THE FIRST ROW, clock pixels etc.
+    mmin, mmax = arr3[1:].min(), arr3[1:].max()
+    # IGNORE THE FIRST ROW, clock pixels etc.
     if subt_ref and lin_scale:
         if mmax > abs(mmin):
             arr3[0, 0] = -mmax
@@ -144,7 +146,6 @@ def arr2im(arr, vmin=0., vmax=10000.0, pwr=1.0, subt_ref=False, lin_scale=True,
             arr3[0, 0] = -mmin
             arr3[0, 1] = -mmax
             mmax = -mmin
-
     arr3 -= mmin
     if mmin < mmax:
         arr3 /= (mmax - mmin)
@@ -185,9 +186,9 @@ def whatfps(fps, crop):
     #define if we are in a predifined crop mode and set max FPS
     cropmode = -1
     fpsmax = max(
-            fps, 600.
-    )  # if unknow configuration, set current fps as max or 600Hz (full frame) as a safeguard. Usually the current frame rate is the max one.
-    for i in range(12):
+            fps, 32000.
+    )  # if unknow configuration, set current fps as max or 3502Hz (full frame) as a safeguard. Usually the current frame rate is the max one.
+    for i in range(setcrops.shape[1]):
         if np.all(crop == setcrops[:, i]):
             cropmode = i
             fpsmax = fpsmaxs[i]
@@ -248,12 +249,12 @@ def updatebiasbpm():
     bpname = conf_dir + "badpixmap%04d_%06d_%03d_%03d_%03d_%03d_%03d.fits" \
              % (fps, etime, ndr, crop[0], crop[2], xsizeim, ysizeim)
     try:
-        badpixmap = pf.getdata(bpname)
+        badpixmap = pf.getdata(bpname).astype(np.bool)
         cam_badpixmap.set_data(badpixmap.astype(np.float32))
         print("badpixmap loaded")
     except:
         bpmhere = False
-        badpixmap = np.ones((ysizeim, xsizeim))
+        badpixmap = np.ones((ysizeim, xsizeim), np.bool)
         print("badpixmap NOT loaded")
     else:
         bpmhere = True
@@ -287,12 +288,12 @@ def make_badpix(bias, filt=3.5):
     ---------
     - filt (default=3.5): filters beyond this many sigmas
     ------------------------------------------------------- '''
-    bpmap = np.ones_like(bias)
+    bpmap = np.ones_like(bias, np.bool)
     rms = np.std(bias)
     mu = np.median(bias)
-    bpmap[bias > mu + filt * rms] = 0.0
-    bpmap[bias < mu - filt * rms] = 0.0
-    return (bpmap)
+    bpmap[bias > mu + filt * rms] = False
+    bpmap[bias < mu - filt * rms] = False
+    return bpmap
 
 
 # ------------------------------------------------------------------
@@ -312,47 +313,43 @@ def whatfilter(reachphoto, slot, block):
 # ------------------------------------------------------------------
 #  Top message
 # ------------------------------------------------------------------
-def whatmsg(pup, reachphoto, ppin, rpin, bpin):
+def whatmsg(reachphoto, gpin, rpin):
     msgtops = [
-            "                ", "     PUPIL      ", "     REACH      ",
-            "REACH PHOTOMETRY", "     GLINT      ", "    APAPANE     "
+            "                ", "     REACH      ", "REACH PHOTOMETRY",
+            "     GLINT      "
     ]
     if reachphoto:
-        msgtop = msgtops[3]
+        msgtop = msgtops[2]
     else:
-        if pup and not rpin:
+        if rpin:
             msgtop = msgtops[1]
-
-        elif rpin:
-            msgtop = msgtops[2]
         else:
-            if ppin:
-                msgtop = msgtops[4]
+            if gpin:
+                msgtop = msgtops[3]
             else:
-                if bpin:
-                    msgtop = msgtops[5]
-                else:
-                    msgtop = msgtops[0]
+                msgtop = msgtops[0]
     return (msgtop)
 
 
 # ------------------------------------------------------------------
 # ------------------------------------------------------------------
 
-hmsg = """PALILA's INSTRUCTIONS
+hmsg = """APAPANE's INSTRUCTIONS
 -------------------
 
 camera controls:
 ---------------
 q           : increase exposure time
 a           : decrease exposure time
+e           : display  CRED1 gain
+w           : increase CRED1 gain
+s           : decrease CRED1 gain
 CTRL+q      : increase number of NDR
 CTRL+a      : decrease number of NDR
 CTRL+SHIFT+<number>: jump to NDR 2**<number>
-CRTL+o      : increase frame rate
+CTRL+o      : increase frame rate
 CTRL+l      : decrease frame rate
 CTRL+h      : hotspotalign
-CTRL+p      : camera to pupil plane/focus plane
 CTRL+i      : REACH mode in/out
 CTRL+b      : take new darks
 CTRL+SHIFT+b: take new dark for current exp
@@ -370,21 +367,18 @@ CTRL+1-6    : change filter wheel slot:
               6. H-band
 CTRL+7      : ircam block
 CTRL+ARROW  : move PSF in focal plane
-CTRL+SHIFT+ARROW  : move PSF in focal plane (more)
 CTRL+ALT+f  : change to full frame
 CTRL+ALT+0-=: change window size:
-  [0]  320 x 256    ( 160-479 x  128-383)  fps = 1500.1 Hz
-  [1]  224 x 188    ( 192-415 x  160-347)  fps = 2050.2 Hz
-  [2]  128 x 128    ( 256-383 x  192-319)  fps = 4500.6 Hz
-  [3]   64 x  64    ( 288-351 x  224-287)  fps = 9203.6 Hz
-  [4]  192 x 192    ( 224-415 x  160-351)  fps = 2200.0 Hz
-  [5]   96 x  72    ( 256-351 x  220-291)  fps = 8002.6 Hz
-  [6]  320 x 256    ( 128-447 x  128-383)  fps = 1500.1 Hz
-  [7]  224 x 188    ( 160-383 x  160-347)  fps = 2050.2 Hz
-  [8]  128 x 128    ( 224-351 x  192-319)  fps = 4500.6 Hz
-  [9]   64 x  64    ( 256-319 x  224-287)  fps = 9203.6 Hz
-  [-]  192 x 192    ( 192-383 x  160-351)  fps = 2200.0 Hz
-  [=]   96 x  72    ( 224-319 x  220-291)  fps = 8002.6 Hz
+        FULL/0:
+        [0] 320 x 256    (   0-319,  0-255 ) fps =  1738.15 Hz
+        [1]  64 x  64    ( 128-191, 96-159 ) fps = 20679.01 Hz
+        [2] 128 x 128    (  96-223, 64-191 ) fps =  7008.36 Hz
+        [3] 160 x 160    (  64-223, 48-207 ) fps =  4926.47 Hz
+        [4] 192 x 192    (  64-255, 32-223 ) fps =  3570.15 Hz
+        [5] 224 x 224    (  32-255, 16-239 ) fps =  2704.52 Hz
+        [6] 256 x 256    (  32-287,  0-255 ) fps =  2117.12 Hz
+        [7] 160 x  80    (  64-223, 88-167 ) fps =  9305.55 Hz
+        [8] 192 x  80    (  64-255, 88-167 ) fps =  8065.61 Hz
 
 display controls:
 ----------------
@@ -396,7 +390,6 @@ o         : bullseye on the PSF
 i         : history of PSF positions
 v         : start/stop accumulating and averaging frames
 g         : seeing measurement (averaging must be on)
-t         : Strehl measurement (averaging must be on)
 z         : zoom/unzoom on the center of the image
 r         : subtract a reference image
 
@@ -405,7 +398,7 @@ mouse controls:
 mouse     : display of the flux under the mouse pointer
 left click: measure distances in mas
 
-ESC       : quit Palila
+ESC       : quit apapane
 
 """
 
@@ -425,30 +418,31 @@ if len(args) >= 2:
 # ------------------------------------------------------------------
 #                access to shared memory structures
 # ------------------------------------------------------------------
-cam = SHM("/milk/shm/palila.im.shm", verbose=False)
+cam = SHM("/milk/shm/apapane.im.shm", verbose=False)
 xsizeim, ysizeim = cam.shape_c
 
-(xsize, ysize) = (320, 256)  #Force size of old palila for the display
+(xsize, ysize) = (320, 256)  #Force size of apapane for the display
 
-cam_dark = cvc.open_shm("palila_dark", dims=(xsizeim, ysizeim), check=True)
-cam_badpixmap = cvc.open_shm("palila_badpixmap", dims=(xsizeim, ysizeim),
+cam_dark = cvc.open_shm("apapane_dark", dims=(xsizeim, ysizeim), check=True)
+cam_badpixmap = cvc.open_shm("apapane_badpixmap", dims=(xsizeim, ysizeim),
                              check=True)
-cam_paused = cvc.open_shm("palila_paused")
-new_dark = cvc.open_shm("palila_newdark")
+
+cam_paused = cvc.open_shm("apapane_paused")
+new_dark = cvc.open_shm("apapane_newdark")
 ircam_synchro = cvc.open_shm("ircam_synchro", dims=(6, 1))
-ircam_retroinj = cvc.open_shm("palila_retroinj", dims=(20, 1))
+ircam_retroinj = cvc.open_shm("ircam_retroinj", dims=(20, 1))
 
 # ------------------------------------------------------------------
 #            Configure communication with SCExAO's redis
 # ------------------------------------------------------------------
 rdb, rdb_alive = cvc.locate_redis_db()
 
-(pup, reachphoto, ppin, rpin, bpin, slot, block, pap, pad, target,
- pdi) = cvc.RDB_pull(rdb, rdb_alive, False, do_defaults=True)
+(pup, reachphoto, gpin, rpin, bpin, slot, block, pap, pad, target,
+ pdi) = cvc.RDB_pull(rdb, rdb_alive, True, do_defaults=True)
 
-pscale = 16.2  #mas per pixel in Palila
+pscale = 16.9  #mas per pixel in Apapane
 
-filename = home + "/conf/palila_aux/hotspots_cor.txt"
+filename = HOME + "/conf/apapane_aux/hotspots_cor.txt"
 cors = [line.rstrip('\n') for line in open(filename)]
 ncor = len(cors)
 cort = np.zeros((2, ncor))
@@ -458,7 +452,7 @@ for i in range(ncor):
     cort[1, i] = float(corparam[3])
 cort /= pscale
 cor = np.array([0, 0])
-if ppin:
+if gpin:
     cor = cort[:, 0]
 
 # ------------------------------------------------------------------
@@ -473,26 +467,30 @@ mycmap = cm.gray
 pygame.display.init()
 pygame.font.init()
 
-#FPSdisp = 20  # frames per second setting
+#FPSdisp = 20  # frames per second setting # Now in argv at top
 fpsClock = pygame.time.Clock()  # start the pygame clock!
 XW, YW = xsize * z1, (ysize + 100) * z1
 
 screen = pygame.display.set_mode((XW, YW), 0, 32)
-pygame.display.set_caption('PALILA camera display!')
+pygame.display.set_caption('APAPANE camera display!')
 
-tmux_ircam_ctrl = tmuxlib.find_or_create_remote(
-        'palila_ctrl', 'scexao@10.20.30.5')  # Control shell
-tmux_ircam = tmuxlib.find_or_create(
-        'palila')  # start a tmux session for messsages
+from swmain.network.pyroclient import connect
+import camstack.pyro_keys
+
+apapane_pyro = connect(
+        camstack.pyro_keys.APAPANE)  # Remote control handle to camera.
+
+tmux_apapane = tmuxlib.find_or_create(
+        'apapane_misc')  # start a tmux session for messsages
 tmux_ircam_synchro = tmuxlib.find_or_create(
         'ircam_synchro')  # start a tmux session for FLC synchro
 
 res = subprocess.check_output("ps aux | grep ircam_synchro", shell=True)
-if bytes(home, 'utf8') + b'/bin/devices/ircam_synchro' not in res:
+if bytes(HOME, 'utf8') + b'/bin/devices/ircam_synchro' not in res:
     tmux_ircam_synchro.send_keys("ircam_synchro")
 
 # ------------------------------------------------------------------
-# -            !!! now we are in business !!!!                     -
+#              !!! now we are in business !!!!
 # ------------------------------------------------------------------
 
 WHITE = (255, 255, 255)
@@ -514,40 +512,37 @@ background = background.convert()
 etimes = np.array([
         10, 20, 50, 100, 200, 500, 1000, 2000, 5000, 10000, 20000, 50000,
         100000, 200000, 500000
-])
+])  # Irrelevant for CRED1
 net = np.size(etimes)
 
-fpss = np.array([10, 20, 50, 100, 200, 500, 1000, 2000, 5000, 10000])
+fpss = np.array([
+        1, 2, 5, 10, 20, 50, 100, 200, 500, 1000, 2000, 5000, 10000, 15000,
+        20000, 25000
+])
 nfps = np.size(fpss)
 fpsmaxs = np.array([
-        1500.2, 2050.2, 4500.6, 9203.6, 2200.0, 8002.6, 1500.2, 2050.2, 4500.6,
-        9203.6, 2200.0, 8002.6
+        3460., 32000., 14331., 9805., 7115., 5390., 4225., 18460., 16020.
 ])
 
-setcrops = np.zeros((4, 12))
-setcrops[:, 0] = [160, 479, 128, 383]
-setcrops[:, 1] = [192, 415, 160, 347]
-setcrops[:, 2] = [256, 383, 192, 319]
-setcrops[:, 3] = [288, 351, 224, 287]
-setcrops[:, 4] = [224, 415, 160, 351]
-setcrops[:, 5] = [256, 351, 220, 291]
-setcrops[:, 6] = [128, 447, 128, 383]
-setcrops[:, 7] = [160, 383, 160, 347]
-setcrops[:, 8] = [224, 351, 192, 319]
-setcrops[:, 9] = [256, 319, 224, 287]
-setcrops[:, 10] = [192, 383, 160, 351]
-setcrops[:, 11] = [224, 319, 220, 291]
+setcrops = np.zeros((4, 9))
+setcrops[:, 0] = [0, 319, 0, 255]
+setcrops[:, 1] = [128, 191, 96, 159]
+setcrops[:, 2] = [96, 223, 64, 191]
+setcrops[:, 3] = [64, 223, 48, 207]
+setcrops[:, 4] = [64, 255, 32, 223]
+setcrops[:, 5] = [32, 255, 16, 239]
+setcrops[:, 6] = [32, 287, 0, 255]
+setcrops[:, 7] = [64, 223, 88, 167]
+setcrops[:, 8] = [64, 255, 88, 167]
 
 ndrs = np.array([1, 2, 4, 8, 16, 32, 64, 128, 255])
 nndr = np.size(ndrs)
 
 # get initial values for expt, fps and ndr
-tmux_ircam_ctrl.send_keys("get_tint()")
-time.sleep(1)
-tmux_ircam_ctrl.send_keys("get_NDR()")
-time.sleep(1)
-tmux_ircam_ctrl.send_keys("get_fps()")
-time.sleep(1)
+apapane_pyro.get_tint()
+apapane_pyro.get_NDR()
+apapane_pyro.get_fps()
+apapane_pyro.get_gain()
 sync_param = ircam_synchro.get_data().astype(np.int)
 lag = 7
 cam_ro = 22 - lag
@@ -563,7 +558,7 @@ else:
     fps = cam.get_fps()
     delay = 0
 ndr = int(cam.get_ndr())
-crop = cam.get_crop().astype(int)
+crop = np.asarray(cam.get_crop()).astype(int)
 etimet = etime * ndr
 
 (fpss2, nfps2, findex) = whatfps(fps, crop)
@@ -583,10 +578,10 @@ font5.set_bold(True)
 xws = xsize * z1
 yws = ysize * z1
 
-path_cartoon = conf_dir + "Palila%d.png" % (z1, )
+path_cartoon = conf_dir + "Apapane%d.png" % (z1, )
 cartoon1 = pygame.image.load(path_cartoon).convert_alpha()
 
-lbl = font1.render("PALILA camera viewer", True, WHITE, BGCOL)
+lbl = font1.render("APAPANE camera viewer", True, WHITE, BGCOL)
 rct = lbl.get_rect()
 rct.center = (110 * z1, 270 * z1)
 screen.blit(lbl, rct)
@@ -607,13 +602,13 @@ rct_info1 = info1.get_rect()
 rct_info1.center = (110 * z1, 305 * z1)
 
 imin, imax = 10000, 10000
-msg2 = ("t = %f" % (etimet))[:8] + (" min,max = %5d,%7d   " % (imin, imax))
+msg2 = ("t = %f" % (etimet))[:8] + (" min,max = %05d,%07d" % (imin, imax))
 info2 = font3.render(msg2, True, FGCOL, BGCOL)
 rct_info2 = info2.get_rect()
 rct_info2.center = (110 * z1, 315 * z1)
 
 xmou, ymou, fmou = 100, 100, 10000
-msg3 = " mouse = %3d,%3d flux = %5d   " % (xmou, ymou, fmou)
+msg3 = " mouse = %3d,%3d flux = %5d    " % (xmou, ymou, fmou)
 info3 = font3.render(msg3, True, FGCOL, BGCOL)
 rct_info3 = info3.get_rect()
 rct_info3.center = (110 * z1, 325 * z1)
@@ -637,11 +632,6 @@ msgsee = "                                     "
 msee = font4.render(msgsee, True, CYAN)
 rct_msee = msee.get_rect()
 rct_msee.center = (xws / 2, 30 * z1)
-
-msgstr = "Strehl = 1.00"
-mstr = font5.render(msgstr, True, CYAN)
-rct_mstr = mstr.get_rect()
-rct_mstr.center = (xws / 2, 30 * z1)
 
 dinfo = font3.render("                     ", True, FGCOL, BGCOL)
 rct_dinfo = dinfo.get_rect()
@@ -682,13 +672,6 @@ sc2 = font4.render(msgsc1, True, CYAN)
 rct_sc2 = sc2.get_rect()
 rct_sc2.bottomleft = (5 * z1 - 4, ysc - ktot)
 
-#REACH Photometry parameters
-dreachphoto = 64. * z1
-xreachphoto = -11.5 * z1
-yreachphoto = -12. * z1
-preachphoto = dreachphoto / 7.
-reachphotoc = preachphoto / 2.
-
 #parallactic angles
 xcpa = xws - 25 * z1
 ycpa = yws - 25 * z1
@@ -718,16 +701,17 @@ rct_zm.topleft = (5 * z1, 5 * z1)
 msgwhl = whatfilter(reachphoto, slot, block)
 wh = font1.render(msgwhl, True, CYAN)
 rct_wh = wh.get_rect()
-rct_wh.topright = (xws - 8 * z1, 5 * z1)
+rct_wh.topright = (xws - 6 * z1, 5 * z1)
 
 #pupil lens
-msgtop = whatmsg(pup, reachphoto, ppin, rpin, bpin)
+msgtop = whatmsg(reachphoto, gpin, rpin)
 topm = font1.render(msgtop, True, CYAN)
 rct_top = topm.get_rect()
 rct_top.midtop = (xws / 2, 5 * z1)
 
 imin, imax = 0, 0
 surf_live = pygame.surface.Surface((xws, yws))
+
 rect1 = surf_live.get_rect()
 rect1.topleft = (0, 0)
 
@@ -770,17 +754,14 @@ logexpt = False  # flag to log the exposure time
 logndr = False  # flag to log the exposure time
 seeing = False
 seeing_plot = False
-strehl = False
-strehl_plot = False
-binary = False
-binary_plot = False
-plot_pa = False  # flag to plot compass roses
+plot_pa = False
 clr_scale = 1  # flag for the display color scale
 shmreload = 0
 keeprpin = False
 wait_for_archive_datatype = False
 
 (badpixmap, bias, bpmhere, biashere) = updatebiasbpm()
+dyn_badpixmap = np.ones_like(badpixmap, np.float32)
 
 ref_im = np.zeros_like(bias) * badpixmap
 
@@ -792,9 +773,6 @@ cnta = 0
 cnti = 0
 timeexpt = []
 timendr = []
-nst = 1
-rm = 15
-a = 128
 
 nhist = 100
 ih = 0
@@ -806,7 +784,6 @@ with open(conf_dir + 'hotspots.txt') as file:
 pos2 = pos[0, :]
 
 # ================================================================================
-# MAIN LOOP
 # ================================================================================
 while True:  # the main game loop
     cnti += 1
@@ -827,10 +804,10 @@ while True:  # the main game loop
                 mycmap = cm.plasma
 
     # ------------------------------------------------------------------
-    # check if camera is paused due to change of window size from other palila
+    # check if camera is paused due to change of window size from other apapane
     campaused = int(cam_paused.get_data())
     while campaused:
-        print("Palila is changing size")
+        print("Apapane is changing size")
         time.sleep(1)
         campaused = int(cam_paused.get_data())
         shmreload = True
@@ -842,13 +819,14 @@ while True:  # the main game loop
         msg = "  !! Acquiring a dark !!  "
         dinfo2 = font3.render(msg, True, BGCOL, SACOL)
         screen.blit(dinfo2, rct_dinfo2)
-        os.system("scexaostatus set darkpalila 'NEW INT DARK    ' 0")
-        os.system("log Palila: Saving current internal dark")
+        tmux_apapane.send_keys(
+                "scexaostatus set darkapapane 'NEW INT DARK    ' 0")
+        tmux_apapane.send_keys("log Apapane: Saving current internal dark")
 
-        print("Palila: dark acquisition.")
+        print("Apapane: acquiring this dark.")
 
-        if not block:
-            os.system("ircam_block")  # blocking the light
+        if not block and bpin:
+            tmux_apapane.send_keys("ircam_block")  # blocking the light
         msgwhl = "     BLOCK      "
         wh = font1.render(msgwhl, True, RED1)
         screen.blit(wh, rct_wh)
@@ -865,14 +843,15 @@ while True:  # the main game loop
         bpname = conf_dir + "badpixmap%04d_%06d_%03d_%03d_%03d_%03d_%03d.fits" \
                  % (fps, etime, ndr, crop[0], crop[2], xsizeim, ysizeim)
         badpixmap = make_badpix(ave_dark)
-        pf.writeto(bpname, badpixmap, overwrite=True)
+        pf.writeto(bpname, badpixmap.astype(np.uint8), overwrite=True)
 
         bias = ave_dark * badpixmap
         time.sleep(0.2)
 
-        os.system("ircam_block")  # blocking the light
-        os.system("scexaostatus set darkpalila 'OFF             ' 1")
-        os.system("log Palila: Done saving current internal dark")
+        tmux_apapane.send_keys("ircam_block")  # blocking the light
+        tmux_apapane.send_keys(
+                "scexaostatus set darkapapane 'OFF             ' 1")
+        tmux_apapane.send_keys("log Apapane: Done saving current internal dark")
         cam_dark.set_data(bias.astype(np.float32))
         cam_badpixmap.set_data(badpixmap.astype(np.float32))
         new_dark.set_data(ONES_NODIM)
@@ -881,24 +860,18 @@ while True:  # the main game loop
     # Relaod shared memory with different size due to change in window size
     if shmreload:
         print("reloading SHM")
-        cam = SHM("/milk/shm/palila.im.shm", verbose=False)
+        cam = SHM("/milk/shm/apapane.im.shm", verbose=False)
         xsizeim, ysizeim = cam.shape_c
-        #(xsizeim, ysizeim) = cam.mtdata['size'][:2]#size[:cam.naxis]
         print("image xsize=%d, ysize=%d" % (xsizeim, ysizeim))
-        #os.system("rm %s/palila_*" % (MILK_SHM_DIR))
         time.sleep(1)
-        cam_dark = cvc.open_shm("palila_dark", dims=(xsizeim, ysizeim),
+        cam_dark = cvc.open_shm("apapane_dark", dims=(xsizeim, ysizeim),
                                 check=True)
-        cam_badpixmap = cvc.open_shm("palila_badpixmap",
+        cam_badpixmap = cvc.open_shm("apapane_badpixmap",
                                      dims=(xsizeim, ysizeim), check=True)
-        #cam_clean = open_shm("palila_clean", dims=(xsizeim, ysizeim), check=True)
-        time.sleep(1)
-        tmux_ircam_ctrl.send_keys("get_tint()")
-        time.sleep(1)
-        tmux_ircam_ctrl.send_keys("get_NDR()")
-        time.sleep(1)
-        tmux_ircam_ctrl.send_keys("get_fps()")
-        time.sleep(1)
+        apapane_pyro.get_tint()
+        apapane_pyro.get_NDR()
+        apapane_pyro.get_fps()
+        apapane_pyro.get_gain()
         shmreload = False
     else:
         # ------------------------------------------------------------------
@@ -910,17 +883,19 @@ while True:  # the main game loop
             fpsn = sync_param[3]
             delay = cam_ro + flc_oft + 3 * lag
         else:
+            # This is where we're going to crash if the FG just stopped / restarted
+            # and we try between the FG starts and the keywords are put in.
             try:
                 etimen = cam.get_expt() * 1e6
-                fpsn = cam.get_fps()
-                delay = 0
             except KeyError:
                 time.sleep(5.)
                 shmreload = True
                 continue
+            fpsn = cam.get_fps()
+            delay = 0
         ndrn = int(cam.get_ndr())
-        cropn = cam.get_crop().astype(int)
-        if ppin:
+        cropn = np.asarray(cam.get_crop()).astype(int)
+        if gpin:
             cor = cort[:, 0]
         else:
             cor = np.array([0, 0])
@@ -936,7 +911,7 @@ while True:  # the main game loop
             ndr = ndrn
             crop = cropn
             etimet = etime * ndr
-            #nindex = np.where(ndrs >= ndr)[0][0] # Commented out for when control server goes to > 255
+            nindex = np.where(ndrs >= ndr)[0][0]
         # ------------------------------------------------------------------
         # read image
         temp, isat = get_img_data(bias, badpixmap, subt_ref, ref_im, lin_scale,
@@ -944,72 +919,26 @@ while True:  # the main game loop
         # ------------------------------------------------------------------
         # averaging
         if average:
-            if rpin:
-                cnta %= 20
             cnta += 1
-            #binary = binary and (cnta == 20)
             if cnta == 1:
                 temp2 = copy.deepcopy(temp)
             else:
                 temp2 *= float(cnta - 1) / float(cnta)
                 temp2 += temp / float(cnta)
-            if seeing or strehl or binary:
-                timestamp = dt.datetime.utcnow().strftime('%Y%m%d')
-                savepath = '/media/data/' + timestamp + '/palilalog/'
-                ospath = os.path.dirname(savepath)
-                if not os.path.exists(ospath):
-                    os.makedirs(ospath)
-                timestamp2 = dt.datetime.utcnow().strftime('%H%M%S')
-                corim = np.median(temp2[:int(ysizeim / 10), :],
-                                  axis=0) + np.median(
-                                          temp2[-int(ysizeim / 10):, :], axis=0)
-                corim /= 2.
-                corim2 = np.tile(corim, (ysizeim, 1))
-                temp2 -= corim2
-                if seeing:
-                    pf.writeto("%s%s_seeing.fits" % (savepath, timestamp2),
-                               temp2, overwrite=True)
-                    se_xstd, se_ystd, se_xc, se_yc, se_theta = impro.calculate_seeing(
-                            temp2)
-                    msgsee = "x = %.2f as, y = %.2f as, t = %d deg" % (
-                            se_ystd * pscale / 1000., se_xstd * pscale / 1000.,
-                            np.rad2deg(se_theta))
-                    os.system(home + "/bin/log SEEING %s: %s" % (slot, msgsee))
-                    seeing = False
-                    seeing_plot = True
-                else:
-                    if "H-band" or "1550" in msgwhl:
-                        flt = "H"
-                    elif "J-band" in msgwhl:
-                        flt = "J"
-                    elif "y-band" in msgwhl:
-                        flt = "y"
-                    else:
-                        flt = ""
-                        print("Strehl calculation not setup for this filter")
-                    if flt != "":
-                        if strehl:
-                            strehlv, dia_core, dia_ring, xoff, yoff, imgstr3 = impro.calculate_strehl(
-                                    temp2, mas_pix=pscale, flt=flt,
-                                    savepath=savepath, timestamp=timestamp2,
-                                    target=target)
-                            msgstr = "Strehl = %.2f" % (strehlv)
-                            os.system(home + "/bin/log STREHL %s: %s" %
-                                      (slot, msgstr))
-                            strehl = False
-                            strehl_plot = True
-                        else:
-                            print("start binary fit")
-                            print(temp2.shape, temp2.max, temp2.min)
-                            posst, xoff, yoff, strehlv, dia_ring, distco, angleco, contrastco = impro.binary_processing(
-                                    temp2, target=target, mas_pix=pscale,
-                                    pad=pad, nst=nst - int(rpin), a=a, rm=rm,
-                                    flt=flt, savepath=savepath,
-                                    timestamp=timestamp2, retroinj=rpin)
-                            os.system(home + "/bin/log BINARY %s: %i-star fit" %
-                                      (slot, nst))
-                            binary = False
-                            binary_plot = True
+            if seeing:
+                #try:
+                pf.writeto("seeing.fits", temp2, overwrite=True)
+                [cx, cy] = impro.centroid(temp2)
+                radc = m.sqrt(np.sum(temp2 > (temp2.max() / 4.)) / m.pi)
+                se_param = impro.fit_TwoD_Gaussian(temp2, cy, cx, radc)
+                seeing = False
+                seeing_plot = True
+                se_ystd = se_param.x_stddev.value
+                se_xstd = se_param.y_stddev.value
+                se_yc = se_param.x_mean.value
+                se_xc = se_param.y_mean.value
+                se_theta = se_param.theta.value
+
         else:
             temp2 = copy.deepcopy(temp)
             cnta = 0
@@ -1017,84 +946,29 @@ while True:  # the main game loop
         imin = np.min(temp2[1:])
         (myim, z3, xmin, xmax, ymin, ymax, xshift,
          yshift) = arr2im(temp2, pwr=pwr0, subt_ref=subt_ref,
-                          lin_scale=lin_scale, pos=pos2)
+                          lin_scale=lin_scale, pos=pos2, pdi=pdi)
         zg = z1 * z2 * z3
         zi = z2 * z3
         pygame.surfarray.blit_array(surf_live, myim)
         screen.blit(surf_live, rect1)
         if average and seeing_plot:
+            msgsee = "x = %.2f as, y = %.2f as, t = %d deg" % (
+                    se_ystd * pscale / 1000. * 2.355,
+                    se_xstd * pscale / 1000. * 2.355, np.rad2deg(se_theta))
             msee = font4.render(msgsee, True, CYAN)
             screen.blit(msee, rct_msee)
             cx = (se_xc + 0.5 - xmin + xshift) * zg
-            cy = (se_yc + 0.5 - ymin + yshift) * zg
-            stdx = se_xstd * zg / 2.
-            stdy = se_ystd * zg / 2.
+            cy = (se_yc - ymin + yshift) * zg
+            stdx = se_xstd * zg * 2.355 / 2.
+            stdy = se_ystd * zg * 2.355 / 2.
             pygame.draw.line(screen, RED1, (cx - stdx * m.cos(se_theta),
-                                            cy + stdx * m.sin(se_theta)),
+                                            cy - stdx * m.sin(se_theta)),
                              (cx + stdx * m.cos(se_theta),
-                              cy - stdx * m.sin(se_theta)), 1)
+                              cy + stdx * m.sin(se_theta)), 1)
             pygame.draw.line(screen, RED1, (cx - stdy * m.sin(se_theta),
-                                            cy - stdy * m.cos(se_theta)),
+                                            cy + stdy * m.cos(se_theta)),
                              (cx + stdy * m.sin(se_theta),
-                              cy + stdy * m.cos(se_theta)), 1)
-        elif average and strehl_plot:
-            mstr = font5.render(msgstr, True, CYAN)
-            screen.blit(mstr, rct_mstr)
-            cx = (int(xsizeim / 2) + 0.5 - xmin + xshift + xoff) * zg
-            cy = (int(ysizeim / 2) + 0.5 - ymin + yshift + yoff) * zg
-            pygame.draw.line(screen, RED1, (cx - bl * zg, cy),
-                             (cx + bl * zg, cy), 1)
-            pygame.draw.line(screen, RED1, (cx, cy - bl * zg),
-                             (cx, cy + bl * zg), 1)
-            pygame.draw.circle(screen, RED1, (int(cx), int(cy)),
-                               int(dia_core / 2 * z2), 1)
-            if dia_ring / 2 * z2 <= yws / 2:
-                pygame.draw.circle(screen, RED1, (int(cx), int(cy)),
-                                   int(dia_ring / 2 * z2), 1)
-        elif average and binary_plot:
-            try:
-                for i in range(nst):
-                    cx = (int(xsizeim / 2) + 0.5 - xmin + xshift + xoff +
-                          posst[i, 0]) * zg
-                    cy = (int(ysizeim / 2) + 0.5 - ymin + yshift + yoff +
-                          posst[i, 1]) * zg
-                    if rpin and i == 1 and cy < yws:
-                        if cy + bl * zg < yws:
-                            pygame.draw.polygon(screen, GREEN,
-                                                [(cx, cy + bl2 * z2),
-                                                 (cx + bl2 * z2 * m.sqrt(3) / 2,
-                                                  cy + bl2 * z2 / 2),
-                                                 (cx + bl2 * z2 * m.sqrt(3) / 2,
-                                                  cy - bl2 * z2 / 2),
-                                                 (cx, cy - bl2 * z2),
-                                                 (cx - bl2 * z2 * m.sqrt(3) / 2,
-                                                  cy - bl2 * z2 / 2),
-                                                 (cx - bl2 * z2 * m.sqrt(3) / 2,
-                                                  cy + bl2 * z2 / 2)], 1)
-                            pygame.draw.line(screen, GREEN, (cx, cy),
-                                             (cx, cy + bl2 * z2), 1)
-                        pygame.draw.line(screen, GREEN, (cx, cy),
-                                         (cx + bl2 * z2 * m.sqrt(3) / 2,
-                                          cy - bl2 * z2 / 2), 1)
-                        pygame.draw.line(screen, GREEN, (cx, cy),
-                                         (cx - bl2 * z2 * m.sqrt(3) / 2,
-                                          cy - bl2 * z2 / 2), 1)
-                    else:
-                        if i == 0 and dia_ring / 2 * z2 <= yws / 2:
-                            pygame.draw.circle(screen, RED1, (int(cx), int(cy)),
-                                               int(dia_ring / 2 * z2), 1)
-                        if cy < yws:
-                            pygame.draw.line(screen, RED1, (cx - bl * zg, cy),
-                                             (cx + bl * zg, cy), 1)
-                        if cy - bl * zg < yws:
-                            pygame.draw.line(screen, RED1, (cx, cy - bl * zg),
-                                             (cx, min(cy + bl * zg, yws - 1)),
-                                             1)
-                        if cx - rm * z2 >= 0 and cx + rm * z2 <= xws and cy - rm * z2 >= 0 and cy + rm * z2 <= yws:
-                            pygame.draw.circle(screen, RED1, (int(cx), int(cy)),
-                                               int(rm * z2), 1)
-            except:
-                print("waiting for new fit")
+                              cy - stdy * m.cos(se_theta)), 1)
 
         # ------------------------------------------------------------------
         # display expt and image information
@@ -1119,14 +993,16 @@ while True:  # the main game loop
         screen.blit(info1, rct_info1)
 
         if etimet < 1e3:
-            msg2 = ("t = %f" % (etimet))[:8] + (" us min,max = %5d,%5d" %
+            msg2 = ("t = %f" % (etimet))[:8] + (" us min,max = %05d,%07d   " %
                                                 (imin, imax))
         elif etimet >= 1e3 and etimet < 1e6:
-            msg2 = ("t = %f" % (etimet / 1.e3))[:8] + (" ms min,max = %5d,%5d" %
-                                                       (imin, imax))
+            msg2 = ("t = %f" %
+                    (etimet / 1.e3))[:8] + (" ms min,max = %05d,%07d   " %
+                                            (imin, imax))
         else:
-            msg2 = ("t = %f" % (etimet / 1.e6))[:8] + (" s  min,max = %5d,%5d" %
-                                                       (imin, imax))
+            msg2 = ("t = %f" %
+                    (etimet / 1.e6))[:8] + (" s  min,max = %05d,%07d   " %
+                                            (imin, imax))
 
         info2 = font3.render(msg2, True, FGCOL, BGCOL)
         rct_info2 = info2.get_rect()
@@ -1161,24 +1037,24 @@ while True:  # the main game loop
                 cx = (int(xsizeim / 2) + cx + 0.5 - xmin + xshift) * zg
                 cy = (int(ysizeim / 2) + cy + 0.5 - ymin + yshift) * zg
                 pygame.draw.line(screen, GREEN, (cxr, cyr),
-                                 (cxr, cyr + bl2 * z2), 1)
+                                 (cxr - bl2 * z2, cyr), 1)
                 pygame.draw.line(screen, GREEN, (cxr, cyr),
-                                 (cxr + bl2 * z2 * m.sqrt(3) / 2,
-                                  cyr - bl2 * z2 / 2), 1)
+                                 (cxr + bl2 * z2 / 2,
+                                  cyr - bl2 * z2 * m.sqrt(3) / 2), 1)
                 pygame.draw.line(screen, GREEN, (cxr, cyr),
-                                 (cxr - bl2 * z2 * m.sqrt(3) / 2,
-                                  cyr - bl2 * z2 / 2), 1)
+                                 (cxr + bl2 * z2 / 2,
+                                  cyr + bl2 * z2 * m.sqrt(3) / 2), 1)
                 pygame.draw.polygon(screen, GREEN, [
-                        (cxr, cyr + bl2 * z2),
-                        (cxr + bl2 * z2 * m.sqrt(3) / 2, cyr + bl2 * z2 / 2),
-                        (cxr + bl2 * z2 * m.sqrt(3) / 2, cyr - bl2 * z2 / 2),
-                        (cxr, cyr - bl2 * z2),
-                        (cxr - bl2 * z2 * m.sqrt(3) / 2, cyr - bl2 * z2 / 2),
-                        (cxr - bl2 * z2 * m.sqrt(3) / 2, cyr + bl2 * z2 / 2)
+                        (cxr + bl2 * z2, cyr),
+                        (cxr + bl2 * z2 / 2, cyr + bl2 * z2 * m.sqrt(3) / 2),
+                        (cxr - bl2 * z2 / 2, cyr + bl2 * z2 * m.sqrt(3) / 2),
+                        (cxr - bl2 * z2, cyr),
+                        (cxr - bl2 * z2 / 2, cyr - bl2 * z2 * m.sqrt(3) / 2),
+                        (cxr + bl2 * z2 / 2, cyr - bl2 * z2 * m.sqrt(3) / 2)
                 ], 1)
 
             else:
-                [cx, cy] = impro.centroid(temp2, method="airy")
+                [cx, cy] = impro.centroid(temp2)
                 if (cx >= 0) and (cx < xsizeim) and (cy >= 0) and (cy
                                                                    < ysizeim):
                     fh = temp2[int(cy), int(cx)]
@@ -1186,18 +1062,18 @@ while True:  # the main game loop
                     info3 = font3.render(msg3, True, FGCOL, BGCOL)
                     screen.blit(info3, rct_info3)
                     cx = (cx + 0.5 - xmin + xshift) * zg
-                    cy = (cy + 0.5 - ymin + yshift) * zg
-            pygame.draw.line(screen, RED1, (cx - bl * zg, cy),
-                             (cx + bl * zg, cy), 1)
-            pygame.draw.line(screen, RED1, (cx, cy - bl * zg),
-                             (cx, cy + bl * zg), 1)
+                    cy = (cy - ymin + yshift) * zg
+            pygame.draw.line(screen, RED1, (cx - bl * z2, cy),
+                             (cx + bl * z2, cy), 1)
+            pygame.draw.line(screen, RED1, (cx, cy - bl * z2),
+                             (cx, cy + bl * z2), 1)
             pygame.draw.circle(screen, RED1, (int(cx), int(cy)), int(bc * z2),
                                1)
 
         # ------------------------------------------------------------------
         # display of position history
         if plot_history:
-            [cx, cy] = impro.centroid(temp2, method="airy")
+            [cx, cy] = impro.centroid(temp2)
             try:
                 fh = temp2[int(cy), int(cx)]
             except:
@@ -1207,7 +1083,7 @@ while True:  # the main game loop
             info3 = font3.render(msg3, True, FGCOL, BGCOL)
             screen.blit(info3, rct_info3)
             coor2[0, ih] = (coor[0, ih] + 0.5 - xmin + xshift) * zg
-            coor2[1, ih] = (coor[1, ih] + 0.5 - ymin + yshift) * zg
+            coor2[1, ih] = (coor[1, ih] - ymin + yshift) * zg
             for ih2 in range(nhist):
                 pygame.draw.line(screen, RED1,
                                  (coor2[0, ih2] - 1, coor2[1, ih2] - 1),
@@ -1218,9 +1094,9 @@ while True:  # the main game loop
             ih += 1
             ih %= nhist
             stds = np.std(coor, axis=1) * pscale
-            cx2 = (np.mean(coor, axis=1)[0] - 320 + crop[0] - pos2[0] -
+            cx2 = (np.mean(coor, axis=1)[0] - 160 + crop[0] - pos2[0] -
                    cor[0]) * pscale
-            cy2 = -(np.mean(coor, axis=1)[1] - 256 + crop[2] - pos2[1] -
+            cy2 = -(np.mean(coor, axis=1)[1] - 128 + crop[2] - pos2[1] -
                     cor[1]) * pscale
             msgcoor = "rms = %.1f mas, %.1f mas, %.1f mas" % (
                     stds[0], stds[1], m.sqrt(np.sum(stds**2) / 2.))
@@ -1269,11 +1145,11 @@ while True:  # the main game loop
             dinfo = font3.render(msg, True, FGCOL, BGCOL)
         screen.blit(dinfo, rct_dinfo)
 
-        if isat > 12000:
+        if isat > 32000:
             msg = "     !!!SATURATION!!!     "
             dinfo2 = font3.render(msg, True, BGCOL, SACOL)
             screen.blit(dinfo2, rct_dinfo2)
-        elif isat > 10000 and isat <= 12000:
+        elif isat > 30000 and isat <= 32000:
             msg = "     !!!NON-LINEAR!!!     "
             dinfo2 = font3.render(msg, True, SACOL, BGCOL)
             screen.blit(dinfo2, rct_dinfo2)
@@ -1302,49 +1178,40 @@ while True:  # the main game loop
         # ------------------------------------------------------------------
         # display the cross
         if plot_cross:
-            if reachphoto:
-                #REACH spot positions
-                for i in range(8):
-                    pygame.draw.circle(screen, GREEN,
-                                       (int(xws / 2 + xreachphoto * zi +
-                                            (i - 3.5) * preachphoto * zi),
-                                        int(yws / 2 + yreachphoto * zi)),
-                                       int(reachphotoc * zi), 1)
+            if pup:
+                #Pupil cross
+                pos2 = pos[1, :]
+                color = GREEN
             else:
-                if pup:
-                    #Pupil cross
-                    pos2 = pos[1, :]
-                    color = GREEN
-                else:
-                    #Focus cross
-                    pos2 = pos[0, :]
-                    color = RED
-                ycross = (256 - crop[2] + pos2[1] + cor[1] - ymin + yshift) * zg
-                xcross = (320 - crop[0] + pos2[0] + cor[0] - xmin + xshift) * zg
-                pygame.draw.line(screen, color, (0, ycross), (xws, ycross), 1)
-                pygame.draw.line(screen, color, (xcross, 0), (xcross, yws), 1)
+                #Focus cross
+                pos2 = pos[0, :]
+                color = RED
+            ycross = (128 - crop[2] + pos2[1] + cor[1] - ymin + yshift) * zg
+            xcross = (160 - crop[0] + pos2[0] + cor[0] - xmin + xshift) * zg
+            pygame.draw.line(screen, color, (0, ycross), (xws, ycross), 1)
+            pygame.draw.line(screen, color, (xcross, 0), (xcross, yws), 1)
 
         if plot_pa:
             pygame.draw.line(screen, RED, (xcpa, ycpa),
+                             (xcpa + 20 * z1 * m.sin(m.radians(pad)),
+                              ycpa - 20 * z1 * m.cos(m.radians(pad))), 1)
+            pygame.draw.line(screen, RED, (xcpa, ycpa),
                              (xcpa - 20 * z1 * m.cos(m.radians(pad)),
                               ycpa - 20 * z1 * m.sin(m.radians(pad))), 1)
-            pygame.draw.line(screen, RED, (xcpa, ycpa),
-                             (xcpa - 20 * z1 * m.sin(m.radians(pad)),
-                              ycpa + 20 * z1 * m.cos(m.radians(pad))), 1)
+            pygame.draw.line(screen, GREEN, (xcpa, ycpa),
+                             (xcpa - 20 * z1 * m.sin(m.radians(pap)),
+                              ycpa + 20 * z1 * m.cos(m.radians(pap))), 1)
             pygame.draw.line(screen, GREEN, (xcpa, ycpa),
                              (xcpa + 20 * z1 * m.cos(m.radians(pap)),
                               ycpa + 20 * z1 * m.sin(m.radians(pap))), 1)
-            pygame.draw.line(screen, GREEN, (xcpa, ycpa),
-                             (xcpa + 20 * z1 * m.sin(m.radians(pap)),
-                              ycpa - 20 * z1 * m.cos(m.radians(pap))), 1)
-            rct_pa1.center = (xcpa - 23 * z1 * m.cos(m.radians(pad)),
+            rct_pa1.center = (xcpa + 23 * z1 * m.sin(m.radians(pad)),
+                              ycpa - 23 * z1 * m.cos(m.radians(pad)))
+            rct_pa2.center = (xcpa - 23 * z1 * m.cos(m.radians(pad)),
                               ycpa - 23 * z1 * m.sin(m.radians(pad)))
-            rct_pa2.center = (xcpa - 23 * z1 * m.sin(m.radians(pad)),
-                              ycpa + 23 * z1 * m.cos(m.radians(pad)))
-            rct_pa3.center = (xcpa + 23 * z1 * m.cos(m.radians(pap)),
+            rct_pa3.center = (xcpa - 23 * z1 * m.sin(m.radians(pap)),
+                              ycpa + 23 * z1 * m.cos(m.radians(pap)))
+            rct_pa4.center = (xcpa + 23 * z1 * m.cos(m.radians(pap)),
                               ycpa + 23 * z1 * m.sin(m.radians(pap)))
-            rct_pa4.center = (xcpa + 23 * z1 * m.sin(m.radians(pap)),
-                              ycpa - 23 * z1 * m.cos(m.radians(pap)))
             screen.blit(pa1, rct_pa1)
             screen.blit(pa2, rct_pa2)
             screen.blit(pa3, rct_pa3)
@@ -1388,21 +1255,26 @@ while True:  # the main game loop
                     exec("rctlines += [rctliner%d]" % i)
 
         # ------------------------------------------------------------------
-        # saving images
+        # saving images ?
+        # Using this construct to find the logger and only itself
+        # Will ret 1 if no processes are found, 0 if the logger is found. Hence the "not"
         '''
-        tmuxon = subprocess.check_output(
-                'ssh scexao-op@localhost "tmux ls" | grep palilalog | awk \'{print $2}\''
-               , shell=True, input=b'')
-        if tmuxon:
+        saving_on = not subprocess.run(
+            'ssh scexaoRTC "ps ax | grep \"milk-logshim apapane\" | grep -v grep"',
+            shell=True,
+            input='',
+            stdout=subprocess.DEVNULL).returncode
+
+        if saving_on:
             saveim = True
             try:  # Assign tmux_ircamlog only if it doesn't exist in the namespace.
                 # This avoid spurious prompts of "duplicate session ircamlog" when logging
-                tmux_ircamlog
+                tmux_apapanelog
             except:
                 # Create a handle to the logging tmux
                 # This allows to get back on track if it already exists when we start palila
-                tmux_ircamlog = tmuxlib.find_or_create_remote(
-                        f'palilalog', 'scexao-op@localhost')
+                tmux_apapanelog = tmuxlib.find_or_create_remote(
+                    'apapane_log', 'scexao@scexaoRTC')
         else:
             saveim = False
         '''
@@ -1411,6 +1283,7 @@ while True:  # the main game loop
             rects = [rect2b, rect2, rct, rct2]
         else:
             rects = []
+
         rects += [
                 rect1, rct_info0, rct_info1, rct_info2, rct_info3, rct_zm,
                 rct_dinfo, rct_dinfo2, rct_sc1, rct_sc2, rct_wh, rct_top
@@ -1427,9 +1300,10 @@ while True:  # the main game loop
             timeexpt = np.append(timeexpt, time.time())
             time.sleep(0.1)
             if timeexpt[-1] - timeexpt[0] > 4:
-                os.system(home +
-                          "/bin/log Palila: changing exposure time to %d" %
-                          etime)
+                tmux_apapane.send_keys(
+                        HOME +
+                        "/bin/log Apapane: changing exposure time to %d" %
+                        etime)
                 timeexpt = []
                 logexpt = False
         if logndr:
@@ -1437,23 +1311,24 @@ while True:  # the main game loop
             timendr = np.append(timendr, time.time())
             time.sleep(0.1)
             if timendr[-1] - timendr[0] > 4:
-                os.system(home +
-                          "/bin/log Palila: changing exposure time to %d" %
-                          etime)
+                tmux_apapane.send_keys(
+                        HOME +
+                        "/bin/log Apapane: changing exposure time to %d" %
+                        etime)
                 timendr = []
                 logndr = False
         if cnti % 20 == 0:
-            if rdb_alive is False:
+            if not rdb_alive:
                 rdb, rdb_alive = cvc.locate_redis_db()
             try:
-                (pup, reachphoto, ppin, rpin, bpin, slot, block, pap, pad,
-                 target, pdi) = cvc.RDB_pull(rdb, rdb_alive, False,
+                (pup, reachphoto, gpin, rpin, bpin, slot, block, pap, pad,
+                 target, pdi) = cvc.RDB_pull(rdb, rdb_alive, True,
                                              do_defaults=rdb_alive)
             except ConnectionError:
                 pass
             msgwhl = whatfilter(reachphoto, slot, block)
             wh = font1.render(msgwhl, True, CYAN)
-            msgtop = whatmsg(pup, reachphoto, ppin, rpin, bpin)
+            msgtop = whatmsg(reachphoto, gpin, rpin)
             topm = font1.render(msgtop, True, CYAN)
 
     # =====================================================================
@@ -1461,7 +1336,7 @@ while True:  # the main game loop
     # =====================================================================
     for event in pygame.event.get():
 
-        # exit Palila
+        # exit Apapane
         #------------------------------------------------------------------
         if event.type == QUIT or (event.type == KEYDOWN and
                                   event.key == K_ESCAPE):
@@ -1472,7 +1347,7 @@ while True:  # the main game loop
             cam_badpixmap.close()
             #cam_clean.close()
             new_dark.close()
-            print("Palila has ended normally.")
+            print("Apapane has ended normally.")
             sys.exit()
 
         elif event.type == KEYDOWN:
@@ -1488,7 +1363,7 @@ while True:  # the main game loop
                     if (nindex < nndr - 1):
                         nindex += 1
                         ndrc = ndrs[nindex]
-                        tmux_ircam_ctrl.send_keys("set_NDR(%d)" % ndrc)
+                        apapane_pyro.set_NDR__oneway(ndrc)
                         time.sleep(1)
                         ndr = cam.get_ndr()
                         etimet = etime * ndr
@@ -1509,12 +1384,6 @@ while True:  # the main game loop
                             etime = sync_param[2]
                             flc_oft = sync_param[4] - lag
                             delay = cam_ro + flc_oft + 3 * lag
-                        else:
-                            tmux_ircam_ctrl.send_keys("set_tint(%f)" %
-                                                      (etimec * 1.e-6, ))
-                            time.sleep(1)
-                            etime = cam.get_expt() * 1e6
-                            delay = 0
                         etimet = etime * ndr
                         (etimes2, net2, tindex) = whatexpt(etime, fps, delay)
                         logexpt = True
@@ -1537,7 +1406,7 @@ while True:  # the main game loop
                     if (nindex > 0):
                         nindex -= 1
                         ndrc = ndrs[nindex]
-                        tmux_ircam_ctrl.send_keys("set_NDR(%d)" % (ndrc, ))
+                        apapane_pyro.set_NDR__oneway(ndrc)
                         time.sleep(1)
                         ndr = cam.get_ndr()
                         etimet = etime * ndr
@@ -1559,8 +1428,7 @@ while True:  # the main game loop
                             flc_oft = sync_param[4] - lag
                             delay = cam_ro + flc_oft + 3 * lag
                         else:
-                            tmux_ircam_ctrl.send_keys("set_tint(%f)" %
-                                                      (etimec * 1.e-6, ))
+                            apapane_pyro.set_tint__oneway(etimec * 1.e-6, )
                             time.sleep(1)
                             etime = cam.get_expt() * 1e6
                             delay = 0
@@ -1576,8 +1444,7 @@ while True:  # the main game loop
                 what_key = DIRECT_NDR_KEYLIST.index(event.key)
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL) and (mmods & KMOD_LSHIFT):
-                    tmux_ircam_ctrl.send_keys("set_NDR(%d)" %
-                                              min(255, 2**what_key))
+                    apapane_pyro.set_NDR__oneway(min(255, 2**what_key))
                     time.sleep(1)
                     ndr = cam.get_ndr()
                     etimet = etime * ndr
@@ -1605,10 +1472,10 @@ while True:  # the main game loop
                             flc_oft = sync_param[4] - lag
                             delay = cam_ro + flc_oft + 3 * lag
                         else:
-                            tmux_ircam_ctrl.send_keys("set_fps(%f)" % (fpsc, ))
+                            apapane_pyro.set_fps__oneway(fpsc)
                             time.sleep(1)
                             fps = cam.get_fps()
-                            tmux_ircam_ctrl.send_keys("get_tint()")
+                            apapane_pyro.get_tint()
                             time.sleep(1)
                             etime = cam.get_expt() * 1e6
                             delay = 0
@@ -1641,10 +1508,10 @@ while True:  # the main game loop
                             flc_oft = sync_param[4] - lag
                             delay = cam_ro + flc_oft + 3 * lag
                         else:
-                            tmux_ircam_ctrl.send_keys("set_fps(%f)" % (fpsc, ))
+                            apapane_pyro.set_fps__oneway(fpsc)
                             time.sleep(1)
                             fps = cam.get_fps()
-                            tmux_ircam_ctrl.send_keys("get_tint()")
+                            apapane_pyro.get_tint()
                             time.sleep(1)
                             etime = cam.get_expt() * 1e6
                             delay = 0
@@ -1661,29 +1528,9 @@ while True:  # the main game loop
             if event.key == K_h:
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL):
-                    tmux_ircam.send_keys("hotspotalign")
+                    tmux_apapane.send_keys("hotspotalign")
                 else:
                     print(hmsg)
-
-            # Camera to Pupil/Focus//Display of parallactic angle
-            #----------------------
-            if event.key == K_p:
-                mmods = pygame.key.get_mods()
-                if (mmods & KMOD_LCTRL):
-                    if pup:
-                        tmux_ircam.send_keys("palila_pup")
-                        if reachphoto:
-                            tmux_ircam.send_keys("palila_pup_fcs pupil &")
-                            tmux_ircam.send_keys("reach_pickoff out &")
-                            tmux_ircam.send_keys("apapane_pickoff out &")
-                        tmux_ircam.send_keys("ircam_fcs palila")
-                    else:
-                        lin_scale = True
-                        tmux_ircam.send_keys("palila_pup")
-                        tmux_ircam.send_keys("palila_pup_fcs pupil &")
-                        tmux_ircam.send_keys("ircam_fcs palila_pup")
-                else:
-                    plot_pa = not plot_pa
 
             # Save new darks for one/all exposure times
             # -----------------------------------------
@@ -1696,17 +1543,15 @@ while True:  # the main game loop
                         msg = "  !! Acquiring a dark !!  "
                         dinfo2 = font3.render(msg, True, BGCOL, SACOL)
                         screen.blit(dinfo2, rct_dinfo2)
-                        os.system(
-                                "scexaostatus set darkpalila 'NEW INT DARK    ' 0"
+                        tmux_apapane.send_keys(
+                                "scexaostatus set darkapapane 'NEW INT DARK    ' 0"
                         )
-                        os.system("log Palila: Saving current internal dark")
+                        tmux_apapane.send_keys(
+                                "log Apapane: Saving current internal dark")
 
-                        print("Palila: dark acquisition.")
-                        if not reachphoto:
-                            if not block:
-                                os.system("ircam_block")  # blocking the light
-                        else:
-                            os.system("PG1_pickoff")
+                        print("Apapane: Acquiring this dark.")
+                        if not block and bpin:
+                            os.system("ircam_block")  # blocking the light
                         msgwhl = "     BLOCK      "
                         wh = font1.render(msgwhl, True, RED1)
                         screen.blit(wh, rct_wh)
@@ -1722,19 +1567,19 @@ while True:  # the main game loop
                         bpname = conf_dir + "badpixmap%04d_%06d_%03d_%03d_%03d_%03d_%03d.fits" \
                                 % (fps, etime, ndr, crop[0], crop[2], xsizeim, ysizeim)
                         badpixmap = make_badpix(ave_dark)
-                        pf.writeto(bpname, badpixmap, overwrite=True)
+                        pf.writeto(bpname, badpixmap.astype(np.uint8),
+                                   overwrite=True)
 
                         bias = ave_dark * badpixmap
                         time.sleep(0.2)
-                        if not reachphoto:
+                        if bpin:
                             os.system("ircam_block")  # blocking the light
-                        else:
-                            os.system("PG1_pickoff")
-                        os.system(
-                                "scexaostatus set darkpalila 'OFF             ' 1"
+                        tmux_apapane.send_keys(
+                                "scexaostatus set darkapapane 'OFF             ' 1"
                         )
-                        os.system(
-                                "log Palila: Done saving current internal dark")
+                        tmux_apapane.send_keys(
+                                "log Apapane: Done saving current internal dark"
+                        )
                         cam_dark.set_data(bias.astype(np.float32))
                         cam_badpixmap.set_data(badpixmap.astype(np.float32))
 
@@ -1744,17 +1589,16 @@ while True:  # the main game loop
                         msg = "  !! Acquiring darks !!   "
                         dinfo2 = font3.render(msg, True, BGCOL, SACOL)
                         screen.blit(dinfo2, rct_dinfo2)
-                        os.system(
-                                "scexaostatus set darkpalila 'ALL INT DARKS   ' 0"
+                        tmux_apapane.send_keys(
+                                "scexaostatus set darkapapane 'ALL INT DARKS   ' 0"
                         )
-                        os.system("log Palila: Saving internal darks")
+                        tmux_apapane.send_keys(
+                                "log Apapane: Saving internal darks")
 
-                        print("Palila: dark acquisition.")
-                        if not reachphoto:
-                            if not block:
-                                os.system("ircam_block")  # blocking the light
-                        else:
-                            os.system("PG1_pickoff")
+                        print("Apapane: Acquiring all darks.")
+                        if bpin:
+                            tmux_apapane.send_keys(
+                                    "ircam_block")  # blocking the light
                         msgwhl = "     BLOCK      "
                         wh = font1.render(msgwhl, True, RED1)
                         screen.blit(wh, rct_wh)
@@ -1774,10 +1618,9 @@ while True:  # the main game loop
                                         np.int)
                                 tint = sync_param[2]
                             else:
-                                tmux_ircam_ctrl.send_keys("set_tint(%f)" %
-                                                          (tint * 1.e-6, ))
+                                apapane_pyro.set_tint__oneway(tint * 1.e-6)
                                 time.sleep(1)
-                                tint = cam.get_expt() * 1e6
+                                tint = cam.get_expt() * 1e3
                             ndark = int(1 * fps /
                                         float(ndr))  # 1s of dark per exposure
                             ave_dark = ave_img_data(ndark, clean=False,
@@ -1789,17 +1632,18 @@ while True:  # the main game loop
                             bpname = conf_dir + "badpixmap%04d_%06d_%03d_%03d_%03d_%03d_%03d.fits" \
                                 % (fps, tint, ndr, crop[0], crop[2], xsizeim, ysizeim)
                             badpixmapi = make_badpix(ave_dark)
-                            pf.writeto(bpname, badpixmapi, overwrite=True)
+                            pf.writeto(bpname, badpixmapi.astype(np.uint8),
+                                       overwrite=True)
 
                             time.sleep(0.2)
-                        if not reachphoto:
-                            os.system("ircam_block")  # opening the shutter
-                        else:
-                            os.system("PG1_pickoff")
-                        os.system(
-                                "scexaostatus set darkpalila 'OFF             ' 1"
+                        if bpin:
+                            tmux_apapane.send_keys(
+                                    "ircam_block")  # opening the shutter
+                        tmux_apapane.send_keys(
+                                "scexaostatus set darkapapane 'OFF             ' 1"
                         )
-                        os.system("log Palila: Done saving internal darks")
+                        tmux_apapane.send_keys(
+                                "log Apapane: Done saving internal darks")
 
                         if not sync_param[0] and sync_param[1]:
                             sync_param[2] = tint
@@ -1807,8 +1651,7 @@ while True:  # the main game loop
                             ircam_synchro.set_data(sync_param.astype(
                                     np.float32))
                         else:
-                            tmux_ircam_ctrl.send_keys("set_tint(%f)" %
-                                                      (tint * 1.e-6, ))
+                            apapane_pyro.set_tint__oneway(tint * 1.e-6)
                         biashere = True
                         bpmhere = True
 
@@ -1816,19 +1659,7 @@ while True:  # the main game loop
             # ---------------------------------------------------
             if event.key == K_r:
                 mmods = pygame.key.get_mods()
-                # Just r
-                if cvc.check_modifiers(mmods):
-                    rname = conf_dir + "ref.fits"
-                    try:
-                        ref_im = pf.getdata(rname) * badpixmap
-                    except:
-                        ref_im = np.zeros((ysizeim, xsizeim))
-                    if ref_im.shape[1] != xsizeim or ref_im.shape[0] != ysizeim:
-                        ref_im = np.zeros((ysizeim, xsizeim))
-                    subt_ref = not subt_ref
-
-                # Ctrl + r
-                elif cvc.check_modifiers(mmods, lc=True):
+                if (mmods & KMOD_LCTRL):
                     msg = "!! Acquiring reference !! "
                     dinfo2 = font3.render(msg, True, BGCOL, SACOL)
                     screen.blit(dinfo2, rct_dinfo2)
@@ -1842,21 +1673,19 @@ while True:  # the main game loop
                     rname = conf_dir + "ref.fits"
                     pf.writeto(rname, ave_ref, overwrite=True)
 
-                # Ctrl + Alt + r
-                elif cvc.check_modifiers(mmods, lc=True, la=True):
-                    tmux_ircam.send_keys(
-                            f"ssh {scxconf.IP_SC5} cam-restartdeps palila")
-                    print("ssh sc5: restarting all Palila auxiliary processes.")
-
-                # Ctrl + Alt + Shift + R
-                elif cvc.check_modifiers(mmods, lc=True, la=True, ls=True):
-                    tmux_ircam.send_keys(
-                            f"ssh {scxconf.IP_SC5} cam-palilastart")
-                    print("ssh sc5: restarting Palila acquisition entirely.")
+                else:
+                    rname = conf_dir + "ref.fits"
+                    try:
+                        ref_im = pf.getdata(rname) * badpixmap
+                    except:
+                        ref_im = np.zeros((ysizeim, xsizeim))
+                    if ref_im.shape[1] != xsizeim or ref_im.shape[0] != ysizeim:
+                        ref_im = np.zeros((ysizeim, xsizeim))
+                    subt_ref = not subt_ref
 
             # Start/stop logging images
             #--------------------------
-            if event.key == K_s:
+            if event.key == K_s:  # Key is S
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL):  # Ctrl + (something) + S
                     saveim = not saveim
@@ -1866,54 +1695,59 @@ while True:  # the main game loop
                             wait_for_archive_datatype = True
                         else:  #  Ctrl + S
                             os.system(
-                                    f"ssh scexao@scexao5 updatekw palila_raw DATA-TYP TEST"
+                                    f"ssh scexao@scexao5 updatekw apapane_raw DATA-TYP TEST"
                             )
                             timestamp = dt.datetime.utcnow().strftime('%Y%m%d')
-                            savepath = '/mnt/tier0/' + timestamp + f'/palila/'
+                            savepath = '/mnt/tier0/' + timestamp + '/apapane/'
+
                             ospath = os.path.dirname(savepath)
 
-                            nimsave = int(min(10000, (20000000 / etimet)))
+                            nimsave = int(min(10000, (10000000 / etimet)))
                             # creating a tmux session for logging
-                            tmux_ircamlog = tmuxlib.find_or_create_remote(
-                                    "palila_log", "scexao@scexaoRTC")
-                            #tmux_ircamlog = tmuxlib.find_or_create("palilalog")
-                            tmux_ircamlog.send_keys(f"mkdir -p {ospath}")
-                            tmux_ircamlog.send_keys(
-                                    "milk-logshim palila %i %s &" %
+                            tmux_apapanelog = tmuxlib.find_or_create_remote(
+                                    "apapane_log", "scexao@scexaoRTC")
+
+                            tmux_apapanelog.send_keys(f"mkdir -p {ospath}")
+                            tmux_apapanelog.send_keys(
+                                    "milk-logshim apapane %i %s &" %
                                     (nimsave, savepath))
-                            os.system("log Palila: start logging images")
+                            os.system("log Apapane: start archiving images")
                             os.system(
-                                    "scexaostatus set logpalila 'LOG NOARCH (RTC)' 3"
+                                    "scexaostatus set logapapane 'LOG NOARCH (RTC)' 3"
                             )
-                    else:
-                        tmux_ircamlog.send_keys(
-                                f"milk-logshimoff palila; sleep 4; milk-logshimkill palila"
+
+                    else:  # We were saving and we're stopping
+                        tmux_apapanelog.send_keys(
+                                "milk-logshimoff apapane; sleep 4; milk-logshimkill apapane"
                         )
-                        os.system("log Palila: stop logging images")
-                        os.system(
-                                "scexaostatus set logpalila 'OFF             ' 1"
+                        #tmux_apapanelog.cmd('kill-session')
+                        tmux_apapane.send_keys(
+                                "log Apapane: stop logging images")
+                        tmux_apapane.send_keys(
+                                "scexaostatus set logapapane 'OFF             ' 1"
                         )
 
-            # Start archiving images
+            # Start archiving images (after prompt to select datatype)
             #--------------------------
             if event.key == K_RETURN and wait_for_archive_datatype:
                 os.system(
-                        f"ssh scexao@scexao5 updatekw palila_raw DATA-TYP {datatyp[idt]}"
+                        f"ssh scexao@scexao5 updatekw apapane_raw DATA-TYP {datatyp[idt]}"
                 )
                 timestamp = dt.datetime.utcnow().strftime('%Y%m%d')
-                savepath = '/media/data/ARCHIVED_DATA/' + timestamp + '/palila/'
+                savepath = '/mnt/tier1/ARCHIVED_DATA/' + timestamp + \
+                           '/apapane/'
                 wait_for_archive_datatype = False
-                ospath = os.path.dirname(savepath)
-                if not os.path.exists(ospath):
-                    os.makedirs(ospath)
-                nimsave = int(min(10000, (20000000 / etimet)))
+                nimsave = int(min(10000, (10000000 / etimet)))
                 # creating a tmux session for logging
-                tmux_ircamlog = tmuxlib.find_or_create_remote(
-                        "palilalog", "scexao@scexaoRTC")
-                tmux_ircamlog.send_keys("milk-logshim palila %i %s &" %
-                                        (nimsave, savepath))
-                os.system("log Palila: start archiving images")
-                os.system("scexaostatus set logpalila 'ARCHIVING (RTC) ' 3")
+                tmux_apapanelog = tmuxlib.find_or_create_remote(
+                        "apapane_log", "scexao@scexaoRTC")
+                ospath = os.path.dirname(savepath)
+
+                tmux_apapanelog.send_keys(f"mkdir -p {ospath}")
+                tmux_apapanelog.send_keys("milk-logshim apapane %i %s &" %
+                                          (nimsave, savepath))
+                os.system("log Apapane: start archiving images")
+                os.system("scexaostatus set logapapane 'ARCHIVING (RTC) ' 3")
 
             # Save an HDR image/Subtract dark
             #--------------------------------
@@ -1926,7 +1760,7 @@ while True:  # the main game loop
                     while ((imax < 4000) & (tindex < net2 - 1)):
                         tindex += 1
                         etime = etimes2[tindex]
-                        cam_cmd("tint %d %d" % (etime), False)
+                        cam_cmd("tint %d" % (etime), False)
                         (badpixmap, bias, bpmhere, biashere) = updatebiasbpm()
                         logexpt = True
                         time.sleep(2)
@@ -1940,7 +1774,7 @@ while True:  # the main game loop
                     while ((isat > 11000) & (tindex > 0)):
                         tindex -= 1
                         etime = etimes2[tindex]
-                        cam_cmd("tint %d %d" % (etime), False)
+                        cam_cmd("tint %d" % (etime), False)
                         (badpixmap, bias, bpmhere, biashere) = updatebiasbpm()
                         logexpt = True
                         time.sleep(2)
@@ -1965,7 +1799,7 @@ while True:  # the main game loop
                             etime2 = copy.deepcopy(etime)
                             tindex += 1
                             etime = etimes2[tindex]
-                            cam_cmd("tint %d %d" % (etime), False)
+                            cam_cmd("tint %d" % (etime), False)
                             (badpixmap, bias, bpmhere,
                              biashere) = updatebiasbpm()
                             logexpt = True
@@ -1986,10 +1820,10 @@ while True:  # the main game loop
 
                     timestamp = dt.datetime.utcnow().strftime('%Y%m%d')
                     timestamp2 = dt.datetime.utcnow().strftime('%H:%M:%S.%f')
-                    savepath = '/media/data/' + timestamp + '/palilalog/'
-                    pf.writeto(savepath + 'palila_hdr_' + timestamp2 + '.fits',
+                    savepath = '/media/data/' + timestamp + '/apapanelog/'
+                    pf.writeto(savepath + 'apapane_hdr_' + timestamp2 + '.fits',
                                hdim / hdim.max(), overwrite=True)
-                    cam_cmd("tint %d %d" % (etimetmp), False)
+                    cam_cmd("tint %d" % (etimetmp), False)
 
                 else:
                     subt_bias = not subt_bias
@@ -2000,22 +1834,13 @@ while True:  # the main game loop
             # Display hotspot crosses
             #------------------------
             if event.key == K_c:
-                mmods = pygame.key.get_mods()
-                if (mmods & KMOD_LCTRL):
-                    if bpin:
-                        os.system("apapane_pickoff out &")
-                        os.system("ircam_fcs palila &")
-                    else:
-                        os.system("apapane_pickoff in &")
-                        os.system("ircam_fcs apapane &")
-                else:
-                    plot_cross = not plot_cross
-                    if plot_cross:
-                        with open(conf_dir + 'hotspots.txt') as file:
-                            pos = np.array([[
-                                    float(digit) for digit in line.split()
-                            ] for line in file])
-                        pos2 = pos[0, :]
+                plot_cross = not plot_cross
+                if plot_cross:
+                    with open(conf_dir + 'hotspots.txt') as file:
+                        pos = np.array([[
+                                float(digit) for digit in line.split()
+                        ] for line in file])
+                    pos2 = pos[0, :]
 
             # Color/grayscale map
             #--------------------
@@ -2030,36 +1855,30 @@ while True:  # the main game loop
                 if (mmods & KMOD_LCTRL):
                     if (mmods & KMOD_LSHIFT):
                         if rpin:
-                            tmux_ircam.send_keys("reach_pickoff out &")
-                            tmux_ircam.send_keys("oap4 onaxis &")
-                            tmux_ircam.send_keys("steering onaxis &")
+                            tmux_apapane.send_keys("reach_pickoff out &")
+                            tmux_apapane.send_keys("oap4 onaxis &")
+                            tmux_apapane.send_keys("steering onaxis &")
                             keeprpin = False
                         else:
-                            tmux_ircam.send_keys("reach_pickoff in &")
-                            tmux_ircam.send_keys("oap4 reach &")
-                            tmux_ircam.send_keys("steering reach &")
+                            tmux_apapane.send_keys("reach_pickoff in &")
+                            tmux_apapane.send_keys("oap4 reach &")
+                            tmux_apapane.send_keys("steering reach &")
                             keeprpin = True
                     else:
                         if reachphoto:
-                            tmux_ircam.send_keys("palila_pup")
-                            tmux_ircam.send_keys("palila_pup_fcs pupil &")
-                            tmux_ircam.send_keys("apapane_pickoff out &")
-                            tmux_ircam.send_keys("ircam_fcs palila &")
+                            tmux_apapane.send_keys("palila_pup")
+                            tmux_apapane.send_keys("palila_pup_fcs pupil &")
+                            tmux_apapane.send_keys("apapane_pickoff out &")
                             if not keeprpin:
-                                tmux_ircam.send_keys("reach_pickoff out &")
-                                tmux_ircam.send_keys("oap4 onaxis &")
-                                tmux_ircam.send_keys("steering onaxis &")
+                                tmux_apapane.send_keys("reach_pickoff out &")
+                                tmux_apapane.send_keys("oap4 onaxis &")
+                                tmux_apapane.send_keys("steering onaxis &")
                         else:
-                            if not pup:
-                                tmux_ircam.send_keys("palila_pup")
-
-                            tmux_ircam.send_keys("palila_pup_fcs reach &")
-                            tmux_ircam.send_keys("apapane_pickoff in &")
-                            tmux_ircam.send_keys("ircam_fcs apapane &")
+                            tmux_apapane.send_keys("apapane_pickoff in &")
                             if not rpin:
-                                tmux_ircam.send_keys("reach_pickoff in &")
-                                tmux_ircam.send_keys("oap4 reach &")
-                                tmux_ircam.send_keys("steering reach &")
+                                tmux_apapane.send_keys("reach_pickoff in &")
+                                tmux_apapane.send_keys("oap4 reach &")
+                                tmux_apapane.send_keys("steering reach &")
                 else:
                     plot_history = not plot_history
 
@@ -2068,11 +1887,6 @@ while True:  # the main game loop
             if event.key == K_v:
                 average = not average
                 seeing_plot = False
-                strehl_plot = False
-                binary_plot = False
-                binary = False
-                nst = 1
-                a = 128
 
             # Start/stop seeing measurement
             #------------------------------
@@ -2081,30 +1895,6 @@ while True:  # the main game loop
                     seeing = True
                 else:
                     seeing = False
-
-            # Start/stop strehl measurement
-            #------------------------------
-            if event.key == K_t:
-                if average:
-                    strehl = True
-                else:
-                    strehl = False
-
-            # Start/stop binary measurement
-            #------------------------------
-            if event.key == K_y:
-                mmods = pygame.key.get_mods()
-                if average:
-                    print("binary fit")
-                    nst += 1
-                    binary = True
-                    if (mmods & KMOD_LSHIFT):
-                        print("big window")
-                        a = 256
-                else:
-                    nst = 1
-                    a = 128
-                    binary = False
 
             # Zoom/unzoom
             #------------
@@ -2124,21 +1914,10 @@ while True:  # the main game loop
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL):
                     if (mmods & KMOD_LALT):
-                        sync_param = ircam_synchro.get_data().astype(np.int)
-                        if not sync_param[0] and not sync_param[1]:
-                            sync_param[2] = etime
-                            sync_param[3] = fps
-                            sync_param[0] = 1
-                            sync_param[1] = 1
-                            ircam_synchro.set_data(sync_param.astype(
-                                    np.float32))
-                            time.sleep(1)
-                        else:
-                            sync_param[0] = 1
-                            sync_param[1] = 0
-                            ircam_synchro.set_data(sync_param.astype(
-                                    np.float32))
-                            time.sleep(1)
+                        # Ctrl+Alt+n -> enable exttrig
+                        apapane_pyro.set_synchro(1)
+                    else:  # Ctrl+n -> disable exttrig
+                        apapane_pyro.set_synchro(0)
 
             # Crop modes and full frame
             #---------------------
@@ -2151,23 +1930,22 @@ while True:  # the main game loop
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL) and (mmods & KMOD_LALT):
                     # Index 12 == Ctrl+alt+f == full
-                    mode_id = (str(what_key), "FULL")[event.key == K_f]
-                    if event.key == K_f and xsizeim == 640 and ysizeim == 512:
+                    mode_id = (what_key, "full")[event.key == K_f]
+                    if event.key == K_f and xsizeim == 320 and ysizeim == 256:
                         # Skip full frame if full frame already
                         print('Camera already in full frame - skipping set_camera_mode()'
                               )
                     else:
                         cam_paused.set_data(ONES_NODIM)
-                        tmux_ircam_ctrl.send_keys("set_camera_mode(%s)" %
-                                                  mode_id)
+                        apapane_pyro.set_camera_mode__oneway(mode_id)
                         # Wait until we're confident the edttake has stopped
                         time.sleep(5.0)
                         # This will return once the SHM has been overwritten...
                         # and hopefully recreated near-immediately after
-                        print('Hi')
+                        print('Hi! Waiting on semaphore til SHM is recreated.')
                         ret = cam.non_block_wait_semaphore()
                         time.sleep(0.1)  # Safe
-                        print('Hi again')
+                        print('Got it.')
                         time.sleep(3.0)  # Safe
                         cam_paused.set_data(ZERO_NODIM)
                         shmreload = True
@@ -2186,15 +1964,15 @@ while True:  # the main game loop
                     else:
                         os.system('ircam_filter %d' % (what_key + 1, ))
 
-            # DM stage/select DATA-TYP for archiving
-            #-----------------------------------------
+            # DM stage
+            #----------
             if event.key == K_UP:
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL):
                     if (mmods & KMOD_LSHIFT):
-                        tmux_ircam.send_keys("dm_stage phi push 1000")
+                        tmux_apapane.send_keys("dm_stage theta push -1000")
                     else:
-                        tmux_ircam.send_keys("dm_stage phi push 100")
+                        tmux_apapane.send_keys("dm_stage theta push -100")
                 else:
                     if wait_for_archive_datatype:
                         idt -= 1
@@ -2204,9 +1982,9 @@ while True:  # the main game loop
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL):
                     if (mmods & KMOD_LSHIFT):
-                        tmux_ircam.send_keys("dm_stage phi push -1000")
+                        tmux_apapane.send_keys("dm_stage theta push +1000")
                     else:
-                        tmux_ircam.send_keys("dm_stage phi push -100")
+                        tmux_apapane.send_keys("dm_stage theta push +100")
                 else:
                     if wait_for_archive_datatype:
                         idt += 1
@@ -2216,21 +1994,59 @@ while True:  # the main game loop
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL):
                     if (mmods & KMOD_LSHIFT):
-                        tmux_ircam.send_keys("dm_stage theta push -1000")
+                        tmux_apapane.send_keys("dm_stage phi push -1000")
                     else:
-                        tmux_ircam.send_keys("dm_stage theta push -100")
+                        tmux_apapane.send_keys("dm_stage phi push -100")
 
             if event.key == K_RIGHT:
                 mmods = pygame.key.get_mods()
                 if (mmods & KMOD_LCTRL):
                     if (mmods & KMOD_LSHIFT):
-                        tmux_ircam.send_keys("dm_stage theta push +1000")
+                        tmux_apapane.send_keys("dm_stage phi push +1000")
                     else:
-                        tmux_ircam.send_keys("dm_stage theta push +100")
+                        tmux_apapane.send_keys("dm_stage phi push +100")
+
+            # Print / incr/decr gain
+            if event.key in [K_w, K_s, K_e]:
+                # No modifiers to avoid conflict with ctrl+S
+                mmods = pygame.key.get_mods()
+                #sleep(.5)
+                gain = float(cam.get_keywords()['DETGAIN'])
+                if not (mmods & KMOD_LSHIFT) and not (mmods & KMOD_LCTRL):
+                    if event.key == K_w:
+                        tar_gain = min(2 * gain, 121)
+                        print(f"set_gain({int(tar_gain)})")
+                        apapane_pyro.set_gain__oneway(int(tar_gain))
+                        time.sleep(.5)
+                    elif event.key == K_s:
+                        if gain == 121:
+                            tar_gain = 64
+                        else:
+                            tar_gain = max(1, gain // 2)
+                        print(f"set_gain({int(tar_gain)})")
+                        apapane_pyro.set_gain__oneway(int(tar_gain))
+                        time.sleep(.5)
+
+                    # All cases, K_w, K_s, K_e
+                    gain = float(cam.get_keywords()['DETGAIN'])
+                    print(f"\n=== Apapane gain: {gain} ===\n")
+
+                if not (mmods & KMOD_LSHIFT) and (
+                        mmods & KMOD_LCTRL) and event.key == K_e:
+                    # Shortcut to 121
+                    print("set_gain(121)")
+                    apapane_pyro.set_gain(121)
+                    time.sleep(.5)
 
     pygame.display.update(rects)
 
+    #pygame.display.flip()
     fpsClock.tick(FPSdisp)
 
 pygame.quit()
 sys.exit()
+
+
+# pyproject.toml entry point.
+def dud_main():
+    pass
